implemented exceptions, exception handler and tested them on Course + added some DTOs for joins

+++ src/main/java/com/example/pizzaspringboot/center_mgmt/service/CourseService.java
@@ -2,10 +2,13 @@ package com.example.pizzaspringboot.center_mgmt.service;
 
 import com.example.pizzaspringboot.center_mgmt.dto.CourseDTO;
 import com.example.pizzaspringboot.center_mgmt.entities.Course;
+import com.example.pizzaspringboot.center_mgmt.exception.AlreadyExistsException;
+import com.example.pizzaspringboot.center_mgmt.exception.NotFoundException;
 import com.example.pizzaspringboot.center_mgmt.repository.CourseRepo;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 import static com.example.pizzaspringboot.center_mgmt.util.EntityMapper.mapDTOToCourse;
@@ -20,15 +23,21 @@ public class CourseService {
     }
 
     // Create
-    public Course createCourse(CourseDTO courseDTO) {
+    public Course createCourse(CourseDTO courseDTO) throws AlreadyExistsException {
+        if (!courseRepo.findByName(courseDTO.getName()).isEmpty()) {
+            throw new AlreadyExistsException("Course with name: " + courseDTO.getName() + " already exists");
+        }
         Course course = mapDTOToCourse(courseDTO);
         return courseRepo.save(course);
-//        return courseRepo.save(courseDTO);
     }
 
     // Read
-    public Course getCourseById(UUID id) {
-        return courseRepo.findById(id).orElse(null);
+    public Course getCourseById(UUID id) throws NotFoundException {
+        Optional<Course> courseOptional = courseRepo.findById(id);
+        if (courseOptional.isEmpty()){
+            throw new NotFoundException("No Course with such id exists");
+        }
+        return courseOptional.get();
     }
 
     public List<Course> getAllCourses() {
@@ -36,19 +45,22 @@ public class CourseService {
     }
 
     // Update
-    public Course updateCourse(CourseDTO courseDTO) {
-        Course course = courseRepo.findById(courseDTO.getId()).orElse(null);
-        if (course != null) {
-            course = mapDTOToCourse(courseDTO);
-            courseRepo.save(course);
+    public Course updateCourse(CourseDTO courseDTO) throws NotFoundException {
+        Optional<Course> courseOptional = courseRepo.findById(courseDTO.getId());
+        if (courseOptional.isEmpty()) {
+            throw new NotFoundException("No Course with such id exists");
         }
-
-        return course;
+        Course savedCourse = courseOptional.get();
+        courseRepo.save(savedCourse);
+        return savedCourse;
     }
 
     // Delete
-    public void deleteCourse(UUID id) {
+    public void deleteCourse(UUID id) throws NotFoundException {
+        Optional<Course> courseOptional = courseRepo.findById(id);
+        if (courseOptional.isEmpty()) {
+            throw new NotFoundException("No Course with such id exists");
+        }
         courseRepo.deleteById(id);
     }
-
 }
